requirements.txt
google-cloud-aiplatform==1.38.0
google-cloud-bigquery==3.13.0
kfp==2.4.0
pandas==2.0.0
scikit-learn==1.3.0
joblib==1.3.0
pyarrow==14.0.0
db-dtypes==1.1.0
